requirements.txt
pillow
exifread
opencv-python
scikit-image
click
rich
binwalk
binutils 
vim-common
imagehash
pytesseract